requirements.txt
fastmcp
pydantic
pyyaml
python-dotenv
httpx
geopy
